+++ src/main/java/kr/hhplus/be/server/application/port/in/PointUseCase.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.application.port.in;
+
+public interface PointUseCase {
+    long charge(String userId, long amount, String idempotencyKey);
+    long getBalance(String userId);
+}
