+++ src/main/java/kr/hhplus/be/server/application/port/in/QueueUseCase.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.application.usecase.queue;
+package kr.hhplus.be.server.application.port.in;
 
 public interface QueueUseCase {
     QueueResult issue(String userId);
